Implemented Paul's review comments.
